All: reduce warnings on VS
It seems MSVC is more strict than gcc.

 * casts from float to uint8_t - bad
 * function args overriding class members - bad
 * `for( auto foo: list)` - you are missing a reference (WOW - this is a bad one!)

I also do not fail the build anymore on MSVC's warnings. This is too intrusive.

I am missing 130  warnings due to stl4043. Ignoring this for now.

+++ src/bitmap.cpp
@@ -76,7 +76,8 @@ auto hslToRGB(const HSL &hsl) -> int32_t {
         b = x;
     }
 
-    return MakeColor((r + m) * 255.0, (g + m) * 255.0, (b + m) * 255.0);
+    return MakeColor(static_cast<uint8_t>((r + m) * 255), static_cast<uint8_t>((g + m) * 255),
+                     static_cast<uint8_t>((b + m) * 255));
 }
 
 auto Darker(uint32_t color, double percentage) -> uint32_t {
